+++ src/core/mod.rs
@@ -1,5 +1,4 @@
 pub(crate) mod blocklist;
-pub(crate) mod blocklist_analysis;
 pub mod config;
 pub mod lookup;
 pub mod network;